+++ src/Type/Definition/Directive.php
@@ -32,7 +32,7 @@ class Directive
     /** @var string|null */
     public $description;
 
-    /** @var FieldArgument[] */
+    /** @var array<int, FieldArgument> */
     public $args = [];
 
     /** @var bool */