+++ src/Type/Definition/UnionType.php
@@ -34,7 +34,7 @@ class UnionType extends Type implements AbstractType, OutputType, CompositeType,
      */
     private $possibleTypeNames;
 
-    /** @var UnionTypeExtensionNode[] */
+    /** @var array<int, UnionTypeExtensionNode> */
     public $extensionASTNodes;
 
     /**
@@ -56,7 +56,7 @@ class UnionType extends Type implements AbstractType, OutputType, CompositeType,
         $this->name              = $config['name'];
         $this->description       = $config['description'] ?? null;
         $this->astNode           = $config['astNode'] ?? null;
-        $this->extensionASTNodes = $config['extensionASTNodes'] ?? null;
+        $this->extensionASTNodes = $config['extensionASTNodes'] ?? [];
         $this->config            = $config;
     }