+++ src/Type/SchemaConfig.php
@@ -38,7 +38,7 @@ class SchemaConfig
     /** @var ObjectType|null */
     public $subscription;
 
-    /** @var Type[]|callable */
+    /** @var array<Type>|(callable(): array<Type>) */
     public $types = [];
 
     /** @var Directive[]|null */
@@ -53,7 +53,7 @@ class SchemaConfig
     /** @var bool */
     public $assumeValid = false;
 
-    /** @var SchemaTypeExtensionNode[] */
+    /** @var array<int, SchemaTypeExtensionNode> */
     public $extensionASTNodes = [];
 
     /**
@@ -207,7 +207,7 @@ class SchemaConfig
     }
 
     /**
-     * @return Type[]|callable
+     * @return array<Type>|(callable(): array<Type>)
      *
      * @api
      */
@@ -217,13 +217,13 @@ class SchemaConfig
     }
 
     /**
-     * @param Type[]|callable $types
+     * @param array<Type>|(callable(): array<Type>) $types
      *
-     * @return SchemaConfig
+     * @return $this
      *
      * @api
      */
-    public function setTypes($types)
+    public function setTypes($types): self
     {
         $this->types = $types;