+++ src/Validator/Rules/QueryComplexity.php
@@ -12,8 +12,10 @@ use GraphQL\Language\AST\FragmentSpreadNode;
 use GraphQL\Language\AST\InlineFragmentNode;
 use GraphQL\Language\AST\Node;
 use GraphQL\Language\AST\NodeKind;
+use GraphQL\Language\AST\NodeList;
 use GraphQL\Language\AST\OperationDefinitionNode;
 use GraphQL\Language\AST\SelectionSetNode;
+use GraphQL\Language\AST\VariableDefinitionNode;
 use GraphQL\Language\Visitor;
 use GraphQL\Language\VisitorOperation;
 use GraphQL\Type\Definition\Directive;
@@ -31,10 +33,10 @@ class QueryComplexity extends QuerySecurityRule
     /** @var int */
     private $maxQueryComplexity;
 
-    /** @var mixed[]|null */
+    /** @var array<string, mixed> */
     private $rawVariableValues = [];
 
-    /** @var ArrayObject */
+    /** @var NodeList<VariableDefinitionNode> */
     private $variableDefs;
 
     /** @var ArrayObject */
@@ -55,7 +57,8 @@ class QueryComplexity extends QuerySecurityRule
     {
         $this->context = $context;
 
-        $this->variableDefs     = new ArrayObject();
+        // @phpstan-ignore-next-line Initializing with an empty array does not set the generic type
+        $this->variableDefs     = new NodeList([]);
         $this->fieldNodeAndDefs = new ArrayObject();
         $this->complexity       = 0;
 
@@ -228,15 +231,18 @@ class QueryComplexity extends QuerySecurityRule
         return false;
     }
 
-    public function getRawVariableValues()
+    /**
+     * @return array<string, mixed>
+     */
+    public function getRawVariableValues(): array
     {
         return $this->rawVariableValues;
     }
 
     /**
-     * @param mixed[]|null $rawVariableValues
+     * @param array<string, mixed>|null $rawVariableValues
      */
-    public function setRawVariableValues(?array $rawVariableValues = null)
+    public function setRawVariableValues(?array $rawVariableValues = null): void
     {
         $this->rawVariableValues = $rawVariableValues ?? [];
     }