+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 #### Unreleased
 - PHP version required: 7.4+
+- Removed deprecated `Type::getInternalTypes()`
 - Remove deprecated `GraphQL::execute()`
+- Remove deprecated `GraphQL::executeAndReturnResult()`
 
 #### 14.6.1
 
@@ -13,6 +15,10 @@ Fix:
 
 Feat:
 - Open ReferenceExecutor for extending
+
+Fix:
+- Ensure properties annotated to hold NodeList are not null
+- Validate that directive argument names do not use reserved or duplicate names
 
 #### 14.5.1